console logging functionality added
