+++ composer.json
@@ -1,0 +1,12 @@
+{
+  "autoload": {
+    "psr-4": {
+      "App\\": "app/",
+      "Tests\\": "tests/"
+    }
+  },
+  "require": {
+    "phpunit/phpunit": "8.*",
+    "kint-php/kint-js": "^2.0"
+  }
+}
